Fix "no such table" error in travis
When doing 'flush', 'makemigrations' and 'migrate', tables might
not be in the database yet, but the modules imported by
apps/curia_vista/urls.py expect them to be there. Don't import
and set these urls if we are doing one of these operations that
might occur when the database has not yet been setup.

This commit should fix the errors in travis and additionally make
it a lot easier to set up a clean build.